separated metrics to own file, removed loading best psnr weights, refactoring
